fix: etcd test v1 and v0
