+++ tests/test_etcd_charm.py
@@ -89,20 +89,20 @@ def generate_mtls_chain(common_name: str) -> tuple[str, str]:
 
 
 @pytest.mark.v0
-def test_deploy_charms_v0(juju_lxd_model: Juju, application_charm_v0: Path) -> None:
+def test_deploy_charms_v0(juju_lxd_model_for_v0: Juju, application_charm_v0_etcd: Path) -> None:
     """Deploy both charms (application and the testing charmed-etcd app) to use in the tests."""
     # Deploy both charms (1 unit for each application to test that later they correctly
     # set data in the relation application databag using only the leader unit).
-    juju_lxd_model.deploy(application_charm_v0, app=REQUIRER_APP_NAME, num_units=1)
-    juju_lxd_model.deploy(ETCD_APP_NAME, channel="3.6/edge", num_units=2)
-    juju_lxd_model.deploy(TLS_NAME, channel="1/edge", config={"ca-common-name": "etcd"})
+    juju_lxd_model_for_v0.deploy(application_charm_v0_etcd, app=REQUIRER_APP_NAME, num_units=1)
+    juju_lxd_model_for_v0.deploy(ETCD_APP_NAME, channel="3.6/edge", num_units=2)
+    juju_lxd_model_for_v0.deploy(TLS_NAME, channel="1/edge", config={"ca-common-name": "etcd"})
 
     # enable TLS and check if the cluster is still accessible
     logger.info("Integrating peer-certificates and client-certificates relations")
-    juju_lxd_model.integrate(f"{ETCD_APP_NAME}:peer-certificates", TLS_NAME)
-    juju_lxd_model.integrate(f"{ETCD_APP_NAME}:client-certificates", TLS_NAME)
-    juju_lxd_model.integrate(REQUIRER_APP_NAME, TLS_NAME)
-    juju_lxd_model.wait(
+    juju_lxd_model_for_v0.integrate(f"{ETCD_APP_NAME}:peer-certificates", TLS_NAME)
+    juju_lxd_model_for_v0.integrate(f"{ETCD_APP_NAME}:client-certificates", TLS_NAME)
+    juju_lxd_model_for_v0.integrate(REQUIRER_APP_NAME, TLS_NAME)
+    juju_lxd_model_for_v0.wait(
         lambda status: apps_active_and_agents_idle(
             status, ETCD_APP_NAME, TLS_NAME, REQUIRER_APP_NAME, idle_period=10
         ),
@@ -112,27 +112,20 @@ def test_deploy_charms_v0(juju_lxd_model: Juju, application_charm_v0: Path) -> N
 
 
 @pytest.mark.v1
-def test_deploy_charms_v1(juju_lxd_model: Juju, application_charm_v1: Path) -> None:
+def test_deploy_charms_v1(juju_lxd_model_for_v1: Juju, application_charm_v1: Path) -> None:
     """Deploy both charms (application and the testing charmed-etcd app) to use in the tests."""
     # Deploy both charms (1 unit for each application to test that later they correctly
     # set data in the relation application databag using only the leader unit).
-    juju_lxd_model.deploy(application_charm_v1, app=REQUIRER_APP_NAME, num_units=1)
-    juju_lxd_model.deploy(ETCD_APP_NAME, channel="3.6/edge", num_units=2)
-    juju_lxd_model.deploy(TLS_NAME, channel="1/edge", config={"ca-common-name": "etcd"})
-    juju_lxd_model.wait(
-        lambda status: apps_active_and_agents_idle(
-            status, ETCD_APP_NAME, TLS_NAME, REQUIRER_APP_NAME, idle_period=10
-        ),
-        timeout=1200,
-        successes=1,
-    )
+    juju_lxd_model_for_v1.deploy(application_charm_v1, app=REQUIRER_APP_NAME, num_units=1)
+    juju_lxd_model_for_v1.deploy(ETCD_APP_NAME, channel="3.6/edge", num_units=2)
+    juju_lxd_model_for_v1.deploy(TLS_NAME, channel="1/edge", config={"ca-common-name": "etcd"})
 
     # enable TLS and check if the cluster is still accessible
     logger.info("Integrating peer-certificates and client-certificates relations")
-    juju_lxd_model.integrate(f"{ETCD_APP_NAME}:peer-certificates", TLS_NAME)
-    juju_lxd_model.integrate(f"{ETCD_APP_NAME}:client-certificates", TLS_NAME)
-    juju_lxd_model.integrate(REQUIRER_APP_NAME, TLS_NAME)
-    juju_lxd_model.wait(
+    juju_lxd_model_for_v1.integrate(f"{ETCD_APP_NAME}:peer-certificates", TLS_NAME)
+    juju_lxd_model_for_v1.integrate(f"{ETCD_APP_NAME}:client-certificates", TLS_NAME)
+    juju_lxd_model_for_v1.integrate(REQUIRER_APP_NAME, TLS_NAME)
+    juju_lxd_model_for_v1.wait(
         lambda status: apps_active_and_agents_idle(
             status, ETCD_APP_NAME, TLS_NAME, REQUIRER_APP_NAME, idle_period=10
         ),
@@ -142,11 +135,36 @@ def test_deploy_charms_v1(juju_lxd_model: Juju, application_charm_v1: Path) -> N
 
 
 @pytest.mark.v0
+def test_relate_client_charm_v0(juju_lxd_model_for_v0: Juju):
+    _relate_client_charm(juju_lxd_model_for_v0)
+
+
 @pytest.mark.v1
-def test_relate_client_charm(
-    juju_lxd_model: Juju,
-    lxd_controller: str,
-) -> None:
+def test_relate_client_charm_v1(juju_lxd_model_for_v1: Juju):
+    _relate_client_charm(juju_lxd_model_for_v1)
+
+
+@pytest.mark.v0
+def test_write_read_with_requirer_v0(juju_lxd_model_for_v0: Juju) -> None:
+    _write_read_with_requirer(juju_lxd_model_for_v0)
+
+
+@pytest.mark.v1
+def test_write_read_with_requirer_v1(juju_lxd_model_for_v1: Juju) -> None:
+    _write_read_with_requirer(juju_lxd_model_for_v1)
+
+
+@pytest.mark.v0
+def test_update_mtls_cert_v0(juju_lxd_model_for_v0: Juju):
+    _update_mtls_cert(juju_lxd_model_for_v0)
+
+
+@pytest.mark.v1
+def test_update_mtls_cert_v1(juju_lxd_model_for_v1: Juju):
+    _update_mtls_cert(juju_lxd_model_for_v1)
+
+
+def _relate_client_charm(juju_lxd_model: Juju) -> None:
     """Test normal client charm relation."""
     juju_lxd_model.integrate(ETCD_APP_NAME, REQUIRER_APP_NAME)
     juju_lxd_model.wait(
@@ -195,9 +213,7 @@ def test_relate_client_charm(
             assert mtls_cert in client_cas, f"mtls cert not in trusted CAs for {unit_name}"
 
 
-@pytest.mark.v0
-@pytest.mark.v1
-async def test_write_read_with_requirer(juju_lxd_model: Juju) -> None:
+def _write_read_with_requirer(juju_lxd_model: Juju) -> None:
     """Test write and read to the key prefix with the requirer charm."""
     requirer_unit = next(iter(juju_lxd_model.status().get_units(REQUIRER_APP_NAME)))
 
@@ -231,9 +247,7 @@ async def test_write_read_with_requirer(juju_lxd_model: Juju) -> None:
         assert lines[1] == TEST_VALUE
 
 
-@pytest.mark.v0
-@pytest.mark.v1
-def test_update_mtls_cert(juju_lxd_model: Juju) -> None:
+def _update_mtls_cert(juju_lxd_model: Juju) -> None:
     """Test updating the common name used by the requirer app."""
     old_mtls_certs = get_requirer_mtls_certificates(juju_lxd_model)
     assert old_mtls_certs, "failed to get the old mtls certs from requirer TLS provider"
