+++ tests/v0/integration/application-charm-etcd-client/requirements.txt
@@ -1,0 +1,4 @@
+ops >= 2.0.0, < 3.0.0
+charmlibs-snap >= 1.0.1
+charmlibs-interfaces-tls-certificates >=1.1.0
+tenacity == 9.1.2
