+++ tests/v0/integration/application-charm/actions.yaml
@@ -43,30 +43,3 @@ delete-relation-field:
 
 set-mtls-cert:
   description: Sets the MTLS cert for the requirer application.
-
-update-mtls-certs:
-  description: Update the mtls certificates
-
-put-etcd:
-  description: Put a key value pair in etcd
-  params:
-    key:
-      description: The key to put in etcd
-      type: string
-    value:
-      description: The value to put in etcd
-      type: string
-
-get-etcd:
-  description: Get a key value pair from etcd
-  params:
-    key:
-      description: The key to get from etcd
-      type: string
-
-get-credentials:
-  description: Action for fetching all available credentials from relations.
-
-get-certificates:
-  description: Get the certificate from the certificate provider
-
