+++ tests/v0/integration/application-charm/metadata.yaml
@@ -39,7 +39,3 @@ requires:
     interface: database_client
   database-with-prefix:
     interface: database_client
-  etcd-client:
-    interface: etcd_client
-  certificates:
-    interface: tls-certificates
