+++ tests/v0/integration/application-charm/requirements.txt
@@ -1,3 +1,1 @@
 ops >= 2.0.0, < 3.0.0
-charmlibs-snap >= 0.8.0
-tenacity == 9.1.2
